+++ lib/src/util/yaml_keys.dart
@@ -51,6 +51,10 @@ final class YamlKeys {
   static const String preventEventParameterDuplicates =
       'prevent_event_parameter_duplicates';
 
+  // Meta config keys
+  static const String autoTrackingCreationDate = 'auto_tracking_creation_date';
+  static const String includeMetaInParameters = 'include_meta_in_parameters';
+
   // Event keys
   static const String eventName = 'event_name';
   static const String description = 'description';